# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DZ Bus Tracker API",
    "description": "Real-time bus tracking for Algeria.\n\n* JSON uses snake_case; timestamps are RFC 3339 UTC.\n* Errors are RFC 9457 problem documents (`application/problem+json`) with a stable `code`, localized in French (default), Arabic or English via `Accept-Language`.\n* Lists use cursor pagination: `?limit=` (1–100) and `?cursor=` from `next_cursor`.\n* Cacheable GETs return a weak `ETag`; send `If-None-Match` to get `304`.\n* Rate limits are reported in `RateLimit-*` headers; `429` carries `Retry-After`.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/lines": {
      "get": {
        "tags": [
          "lines"
        ],
        "summary": "List lines (newest first). Active lines only, unless the caller administers lines.",
        "operationId": "list_lines",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Part of the code or name (2–100 characters, case-insensitive).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stop_id",
            "in": "query",
            "description": "Only lines serving this stop.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "description": "Activity filter, for line administrators only (`403` otherwise).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of lines",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_LineDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "401": {
            "description": "`is_active` filter without being signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "`is_active` filter without `line:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid filters (e.g. `q` shorter than 2 characters)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "lines"
        ],
        "summary": "Create a line (audited).",
        "operationId": "create_line",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Line created (`Location` points to it)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `line:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Another line has this code (`line_code_taken`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/lines/{id}": {
      "get": {
        "tags": [
          "lines"
        ],
        "summary": "One line, with its number of stops and whether it has a route.",
        "operationId": "get_line",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The line",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "404": {
            "description": "No such (visible) line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "lines"
        ],
        "summary": "Delete a line with its stop list and schedules (audited); refused while buses are assigned\nto it.",
        "operationId": "delete_line",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `line:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Buses are assigned to the line (`line_in_use`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "lines"
        ],
        "summary": "Update a line partially (audited). The code cannot change; `is_active` activates or\ndeactivates the line.",
        "operationId": "update_line",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated line",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `line:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields (`code` is `unknown_field`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/lines/{id}/route": {
      "get": {
        "tags": [
          "lines"
        ],
        "summary": "The route of a line as a GeoJSON `LineString`.",
        "operationId": "get_line_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (If-None-Match)"
          },
          "404": {
            "description": "No such (visible) line, or no route set",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "lines"
        ],
        "summary": "Set the route of a line (audited). Replaces the legacy route segments, which any user could\nwrite (L-04).",
        "operationId": "set_line_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `line:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Not a valid `LineString`: 2–10 000 positions in range, no repeated consecutive position",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "lines"
        ],
        "summary": "Remove the route of a line (idempotent, audited).",
        "operationId": "delete_line_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The line has no route any more"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Missing `line:write`",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such line",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDocument": {
        "type": "object",
        "description": "RFC 9457 problem document returned by every error response.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            }
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "URI identifying the problem type (`urn:dzbus:problem:<code>`)."
          }
        }
      },
...
      "UpdateLineRequest": {
        "type": "object",
        "description": "Changes to a line (`code` is immutable). `null` clears `frequency_minutes` and `fare_dza`;\n`is_active` replaces the legacy activate/deactivate actions.",
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fare_dza": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "frequency_minutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Machine-to-machine key (`dzk_…`) with explicit scopes."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from /auth/login or /auth/refresh (EdDSA)."
      }
    }
  },
...
}